Refactor: footer의 위치를 Layout 폴더로 이동했다
